+++ src/server/rest/v1/router/api/BillingRouter.ts
@@ -2,7 +2,6 @@ import { ServerAction, ServerRoute } from '../../../../../types/Server';
 import express, { NextFunction, Request, Response } from 'express';
 
 import BillingService from '../../service/BillingService';
-import BillingSettingService from '../../service/BillingSettingService';
 import RouterUtils from '../RouterUtils';
 
 export default class BillingRouter {
@@ -40,37 +39,37 @@ export default class BillingRouter {
 
   protected buildRouteBillingSettings(): void {
     this.router.get(`/${ServerRoute.REST_BILLING_SETTINGS}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleGetBillingSettings.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleGetBillingSettings.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
 
   protected buildRouteBillingSetting(): void {
     this.router.get(`/${ServerRoute.REST_BILLING_SETTING}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleGetBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleGetBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
 
   protected buildRouteUpdateBillingSetting(): void {
     this.router.put(`/${ServerRoute.REST_BILLING_SETTING}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleUpdateBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleUpdateBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
 
   protected buildRouteCheckBillingSetting(): void {
     this.router.post(`/${ServerRoute.REST_BILLING_SETTING_CHECK}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleCheckBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleCheckBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
 
   protected buildRouteActivateBillingSetting(): void {
     this.router.post(`/${ServerRoute.REST_BILLING_SETTING_ACTIVATE}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleActivateBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleActivateBillingSetting.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
 
   protected buildRouteCheckBillingSettingConnection(): void {
     this.router.post(`/${ServerRoute.REST_BILLING_SETTING_CHECK_CONNECTION}`, (req: Request, res: Response, next: NextFunction) => {
-      void RouterUtils.handleServerAction(BillingSettingService.handleCheckBillingSettingConnection.bind(this), ServerAction.SETTINGS, req, res, next);
+      void RouterUtils.handleServerAction(BillingService.handleCheckBillingSettingConnection.bind(this), ServerAction.SETTINGS, req, res, next);
     });
   }
